permission authorize

+++ routes/web.php
@@ -21,40 +21,40 @@ use Illuminate\Support\Facades\Route;
 
 Route::middleware('auth')->group(function(){
     Route::get('/', [UserController::class, 'index']);//index page
-    Route::get('/profile/{name}', [UserController::class, 'profile']);//profile page
+    Route::get('/profile/{name}', [UserController::class, 'profile'])->middleware('permission:view profile');//profile page
 
     //employee page
-    Route::resource('employee', EmployeeController::class);//employee table
-    Route::get('employee/create', [EmployeeController::class, 'createView']);//create employee
-    Route::post('employee/create', [EmployeeController::class, 'store']);//store employee
-    Route::get('/employee/{id}/info', [EmployeeController::class, 'show']);  // employee detail
-    Route::get('/employee/{id}/edit', [EmployeeController::class, 'edit']);  // edit employee
-    Route::patch('/employee/{id}/update', [EmployeeController::class, 'update']);  // update employee
-    Route::delete('/employee/{id}/delete', [EmployeeController::class, 'delete']);  // delete employee
+    Route::resource('employee', EmployeeController::class)->middleware('permission:view employees');//employee table
+    Route::get('employee/create', [EmployeeController::class, 'createView'])->middleware('permission:create employees');//create employee
+    Route::post('employee/create', [EmployeeController::class, 'store'])->middleware('permission:create employees');//store employee
+    Route::get('/employee/{id}/info', [EmployeeController::class, 'show']);// employee detail
+    Route::get('/employee/{id}/edit', [EmployeeController::class, 'edit'])->middleware('permission:edit employees');  // edit employee
+    Route::patch('/employee/{id}/update', [EmployeeController::class, 'update'])->middleware('permission:edit employees');  // update employee
+    Route::delete('/employee/{id}/delete', [EmployeeController::class, 'delete'])->middleware('permission:delete employees'); // delete employee
 
     //department page
-    Route::resource('department', DepartmentController::class);//department table
-    Route::get('department/create', [DepartmentController::class, 'createView']);//create department
-    Route::post('department/create', [DepartmentController::class, 'store']);//store department
-    Route::get('/department/{id}/edit', [DepartmentController::class, 'edit']);  // edit department
-    Route::patch('/department/{id}/update', [DepartmentController::class, 'update']);  // update department
-    Route::delete('/department/{id}/delete', [DepartmentController::class, 'delete']);  // delete department
+    Route::resource('department', DepartmentController::class)->middleware('permission:view departments');//department table
+    Route::get('department/create', [DepartmentController::class, 'createView'])->middleware('permission:create departments');//create department
+    Route::post('department/create', [DepartmentController::class, 'store'])->middleware('permission:create departments');//store department
+    Route::get('/department/{id}/edit', [DepartmentController::class, 'edit'])->middleware('permission:edit departments');  // edit department
+    Route::patch('/department/{id}/update', [DepartmentController::class, 'update'])->middleware('permission:edit departments');  // update department
+    Route::delete('/department/{id}/delete', [DepartmentController::class, 'delete'])->middleware('permission:delete departments');  // delete department
 
     //role page
-    Route::resource('role', RoleController::class);//role table
-    Route::get('role/create', [RoleController::class, 'createView']);//create role
-    Route::post('role/create', [RoleController::class, 'store']);//store role
-    Route::get('/role/{id}/edit', [RoleController::class, 'edit']);  // edit role
-    Route::patch('/role/{id}/update', [RoleController::class, 'update']);  // update role
-    Route::delete('/role/{id}/delete', [RoleController::class, 'delete']);  // delete role
+    Route::resource('role', RoleController::class)->middleware('permission:view roles');//role table
+    Route::get('role/create', [RoleController::class, 'createView'])->middleware('permission:create roles');//create role
+    Route::post('role/create', [RoleController::class, 'store'])->middleware('permission:create roles');//store role
+    Route::get('/role/{id}/edit', [RoleController::class, 'edit'])->middleware('permission:edit roles');  // edit role
+    Route::patch('/role/{id}/update', [RoleController::class, 'update'])->middleware('permission:edit roles');  // update role
+    Route::delete('/role/{id}/delete', [RoleController::class, 'delete'])->middleware('permission:delete roles');  // delete role
 
     //permission page
-    Route::resource('permission', PermissionController::class);//permission table
-    Route::get('permission/create', [PermissionController::class, 'createView']);//create permission
-    Route::post('permission/create', [PermissionController::class, 'store']);//store permission
-    Route::get('/permission/{id}/edit', [PermissionController::class, 'edit']);  // edit permission
-    Route::patch('/permission/{id}/update', [PermissionController::class, 'update']);  // update permission
-    Route::delete('/permission/{id}/delete', [PermissionController::class, 'delete']);  // delete role
+    Route::resource('permission', PermissionController::class)->middleware('permission:view permissions');//permission table
+    Route::get('permission/create', [PermissionController::class, 'createView'])->middleware('permission:create permissions');//create permission
+    Route::post('permission/create', [PermissionController::class, 'store'])->middleware('permission:create permissions');//store permission
+    Route::get('/permission/{id}/edit', [PermissionController::class, 'edit'])->middleware('permission:edit permissions');  // edit permission
+    Route::patch('/permission/{id}/update', [PermissionController::class, 'update'])->middleware('permission:edit permissions');  // update permission
+    Route::delete('/permission/{id}/delete', [PermissionController::class, 'delete'])->middleware('permission:delete permissions');  // delete role
 
     Route::post('/logout', [AuthUserController::class, 'logout']);// user logout
 });
